Added Biomes o Plenty.
Why does Curse drop the mod if it's not in the mod folder? Surely the record of it in the json data file would trigger it to just download the file? Now I have to distribute the mods, which was one of the main things I was trying to avoid by using the Curse launcher.